beacon2-pi-api/requirements.txt
# I noticed a misbehaviour with 3.7.2 regarding cookie storage for raised HTTP response
# That's the case for the HTTPFound redirections
# It seems the cookies are not sent in the response, and therefore, the browser does not store them
# Obviously, sessions do not work then, and Openid Connect auth.py code is broken
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiohttp-middlewares==2.4.0
aiohttp-jinja2==1.5
aiohttp-session==2.9.0
asyncpg==0.24.0
cryptography==41.0.5
PyJWT==2.7.0
python-dotenv==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
pyyaml==6.0.1
dataclasses-json==0.5.6
pymongo[srv]==4.0.1
pyopenssl==23.3.0
coverage==7.6.0
requests==2.25.1
pydantic==2.6.2
StrEnum==0.4.7
pyhumps==3.8.0
paramiko==2.7.2
six==1.16.0
progressbar==2.5
tqdm==4.62.3
obonet==0.3.1